pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "oh-my-codex"
version = "0.3.0"
description = "Multi-agent orchestration for OpenAI Codex CLI - Turn Codex into a team of AI agents"
readme = "README.md"
license = {text = "MIT"}
authors = [
    {name = "junghwaYang", email = "junghwa@example.com"}
]
keywords = ["codex", "openai", "agents", "orchestration", "cli", "ai", "multi-agent", "gpt"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Software Development :: Code Generators",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
requires-python = ">=3.10"
dependencies = [
    "openai>=1.0.0",
    "pyyaml>=6.0.0",
]

[project.optional-dependencies]
full = [
    "openai-agents>=0.0.1",
]
analytics = [
    "pandas>=2.0.0",
]
dev = [
    "pytest>=7.0.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
]

[project.scripts]
omx = "orchestrator.cli:main"
omx-setup = "orchestrator.setup:main"

[project.urls]
Homepage = "https://github.com/junghwaYang/oh-my-codex"
Repository = "https://github.com/junghwaYang/oh-my-codex"
Documentation = "https://github.com/junghwaYang/oh-my-codex#readme"

[tool.setuptools.packages.find]
where = ["."]
include = ["orchestrator*"]

[tool.setuptools.package-data]
"*" = ["*.md", "*.toml"]

[tool.black]
line-length = 100
target-version = ["py310", "py311", "py312"]

[tool.ruff]
line-length = 100
select = ["E", "F", "W", "I"]
ignore = ["E501"]
